[MIG] mail_quoted_reply: Migration to 18.0
